organize-collection-view: Don't mark local collections as remote
We should hide the details cell renderer if we failed to find the
collection or if it is local. We were only hiding it in the first case.
It meant that if a remote collection was followed by a local one, then
the remote's source name would leak into the local collection's row.

Fall out from e8cd66f34e36660a797f347946bd43fc4472bbe5

https://bugzilla.gnome.org/show_bug.cgi?id=768500