base-item: Avoid pointing to invalid memory
When setting a new default GAppInfo we should not only free the old
name, but also NULLify it. Otherwise, if were passed a NULL GAppInfo,
then the pointers to the name and the object would be out of sync. One
would be non-NULL and the other NULL. We would also confuse the NULL
checks and risk accessing invalid memory.