application: Remove any future timeouts before refreshing a miner
We call RefreshDB on the miners at certain times:
 - at startup
 - at 1 minute intervals thereafter
 - when there is a change in the available Sources

In the last case, we might end up calling RefreshDB when it has also
been scheduled to run at some point in the future. This violates an
assertion in photos_application_refresh_db that protects against it.
Since a change in the available Sources indicates an explicit user
action, it will be nice to immediately react to it. Let's remove the
timeout and then do so.