base-item: Close the GOutputStream before copying the metadata
We haven't encountered any corrupted files, but it looks like a good
idea to close the GOutputStream used by GdkPixbuf to copy the pixel
data before using GExiv2 to copy the metadata. This does change the
SHA256 checksum of the output file, but I can't spot any breakage. If
anything, things should be better now.