item-manager: Add photos_item_manager_(un)hide_item
Right now PhotosItemManager doesn't have any idea about items that were
removed from the UI as a precursor to an actual deletion that will
only happen if the user doesn't undo it. This means that it cannot
properly handle TrackerChangeMonitor::changes-pending for such items.
Secondly, when we change PhotosItemManager to have multiple child
managers, we will need to track the modes for each hidden item so that
we can restore them to their appropriate child managers. Making
ItemManager aware of hidden items will help with that.

https://bugzilla.gnome.org/show_bug.cgi?id=770157