Split ItemManager to have a separate child manager for each mode
So far, we had separate TrackerControllers for each mode (except EDIT
and PREVIEW). These would issue SPARQL queries to match the purpose of
each mode. The results of these queries would go into a single bucket
inside ItemManager, only to be re-filtered by the respective
ViewModels. This is bad. Since we already issue separate SPARQL
queries for each mode, we should keep the results in separate buckets
instead of mixing them up.

ItemManager now has separate BaseManagers for each mode. The
implementation of the ItemManager API has been adjusted to relay the
calls to the child BaseManagers. The 0th child BaseManager is
special-cased to be the union of all the other children.

The next step is to change ViewModel to mirror the content of the
mode-specific child BaseManager, instead of the entire ItemManager.
That will help simplify a lot of the complex (re-)filtering code
inside ViewModel.

Some changes by Debarshi Ray.

https://bugzilla.gnome.org/show_bug.cgi?id=770157