Don't add a collection while we are browsing another one
Since viewing a list of collections and browsing a specific collection
share the same mode, we need to be careful that we don't insert a
collection while we are browsing another.

This can happen if a collection was created out of band (eg., by the
miners), or if one was activated via the search provider. In the first
case, we can ignore it because it is irrelevant to the current state
of the application. In the second case, we should force it to come out
of the current collection and then insert it.

https://bugzilla.gnome.org/show_bug.cgi?id=770342