Remove the mode-specific filtering from PhotosViewModel
Now that ItemManager has separate child BaseManagers for each mode, we
can greatly simplify ViewModel by removing the mode-specific filtering
logic. This was needed earlier when we had one single global bucket
inside ItemManager where the results of the mode-specific queries, and
we had to filter the contents of the bucket for each ViewModel. Not any
more.

From now on, ViewModel will simply mirror the contents of each
mode-specific child BaseManager.

The code to move a BaseItem from one mode to another, when any of its
attributes change (eg., it gets favorited), now resides inside
ItemManager.

https://bugzilla.gnome.org/show_bug.cgi?id=770342