item-manager: Fix logic to check if a cursor is a collection
Some collections might not have a nao:identifier. eg., the built-in
nfo:image-category-screenshot. Instead, we should check the rdf:type
for nfo:DataContainer. This is also what PhotosBaseItem does.

This wasn't a problem so far because we were miraculously not hitting
this code path for such collections. Things are going to change in
subsequent commits when we sprinkle some assertions around the code.

https://bugzilla.gnome.org/show_bug.cgi?id=770342