preview-view, preview-nav-buttons: Don't walk past the edges
It is possible to walk past the edges when navigating very fast with
the keyboard. This is because the app.load-next and app.load-previous
GActions are implemented by both PreviewView and PreviewNavButtons. One
of them swaps out the ImageView and the other triggers the loading. It
will be more robust if only one of them implements the GActions and
drives the other as and when needed.

This partially reverts commit bcdc22c208acabce039ec9491cf1aa5f9eeaca85
by re-introducing the signals in PreviewNavButtons.