Revert "view-container: Assert that selection is restricted ..."
Avoid triggering this assertion while searching:
  CRITICAL **: photos_view_container_view_selection_changed: assertion
    'self->mode == mode' failed

GdMainView emits view-selection-changed even when the underlying
GtkTreeModel is cleared without anything being selected. This happens
when we search. The ViewModel for the SEARCH mode is cleared when
TrackerSearchController issues the search query. Unfortunately, due to
the ordering of the SearchController::search-string-changed handlers,
this happens before the application has switched to the SEARCH mode and
the assertion is triggered.

This reverts commit a243a9110ce6edb1a786d502ced2ec21594acc2b.