item-manager: Add photos_item_manager_clear
Encouraging the TrackerControllers to take a peek at the child
BaseManagers doesn't sound like a good idea. Plus, it is slightly more
efficient to update the 0th child if the ItemManager is in control of
the operation - we can iterate over the, hopefully, smaller child
BaseManager instead of the union of all children.

https://bugzilla.gnome.org/show_bug.cgi?id=770342