base-item: Add a metadata_add_shared virtual method
Once an item has been shared, we want to add some metadata identifying
the shared copy inside the original. This will let us re-create the
nie:links and nie:relatedTo properties in Tracker if we ever lose the
database, or if the user changes computers.

Different BaseItem sub-classes can implement this virtual method to
have different mechanisms for adding this metadata. For LocalItems, we
are going to embed it into the XMP metadata. The default implementation
is a no-op.

https://bugzilla.gnome.org/show_bug.cgi?id=770267