item-manager: Add photos_item_manager_wait_for_changes_async
This is meant to let us work around the URN possibly changing due to
an atomic file update. It will wait for any relevant changes and return
the new URN, if any. This is only needed for LocalItems since the
other BaseItem sub-classes are not backed by a file on the local hard
disk.

https://bugzilla.gnome.org/show_bug.cgi?id=770267
https://bugzilla.gnome.org/show_bug.cgi?id=771042