build: Disable rdtscp by default
... because it is only available in various Core i[357] CPUs and is
not portable across all x86_64 versions. It is not prudent to
automatically enable it during the build because a newer build machine
will cause the application to crash with a SIGILL when run on
something older.

This is also a good time to switch to GB_ENABLE_RDTSCP.

Original patch from Christian Hergert for gnome-builder.

https://bugzilla.gnome.org/show_bug.cgi?id=771995