application: Remove the pointer for the object method signal handler
This is a final class, so there wasn't going to be any sub-class that
would want to override the default signal handler. We might still want
a default handler for PhotosApplication itself, but we can achieve the
same effect with a simple private function. It is a small price to pay
for being able to use G_DECLARE_FINAL_TYPE.

https://bugzilla.gnome.org/show_bug.cgi?id=763712