item-manager: Avoid CRITICALs when favouriting contents of collections
Favouriting an item inside a collection leads to BaseItem::info-updated
being emitted for both the item and the collection. When this is done
while browsing the collection, the collection item is not present
inside ItemManager because we have cleared out all the collections so
that we can share the same mode for listing collections and browsing
them. This leads to:
  CRITICAL **: photos_item_manager_info_updated: assertion
    'updated_item == item' failed

Let's work around this until we have a separate mode for browsing
collections.

Fallout from 5497a7c67c13509da605bd51de1098101c7a40b7

https://bugzilla.gnome.org/show_bug.cgi?id=770342