Turn PhotosBaseManager and PhotosItemManager into GListModels
This is primarily meant to replace PhotosViewModel and our
GtkIconView-based views with ones based on GtkFlowBox.

As a positive side-effect, photos_base_manager_get_objects has been
replaced with the GListModel API for iterating over the contents of a
PhotosBaseManager. Unlike a GHashTable, GListModel returns the objects
in a well-defined order which is better for displaying them in the UI,
and we don't have to expose any internal implementation details.

Based on ideas from Alessandro Bono.

https://bugzilla.gnome.org/show_bug.cgi?id=774191