preview-nav-buttons: Fix the logic to avoid long presses
GtkGestureMultiPress::released is emitted, and it is meant to be, even
after we denied all the sequences in GtkGestureLongPress::pressed.
Therefore, we were toggling the overlay even on a long press.

We are supposed to check the state of the sequence ourselves, and to
do that we need access to the GdkEventSequence object. Hence, we now
connect to GtkGesture::end instead of GtkGestureMultiPress::released.