embed, main-toolbar: Don't use ViewModel to enable/disable the buttons
Taking the ViewContainer's ViewModel and passing it to the MainToolbar
was an ugly thing to do in the first place. Now that every mode has a
separate PhotosBaseManager, which acts as a GListModel, we can do this
in a better.

This further reduces our use of PhotosViewModel and clears the way for
the GtkFlowBox port.

https://bugzilla.gnome.org/show_bug.cgi?id=775562