main-toolbar: Avoid CRITICALs when activating an item
There are no child BaseManagers for the EDIT and PREVIEW modes. Hence
it lead to:
  CRITICAL **: photos_item_manager_get_for_mode: assertion
    'mode != PHOTOS_WINDOW_MODE_PREVIEW' failed

This is also a good time to use a switch statement instead of if,
because -Wswitch-enum will warn us if there are any additions to
PhotosWindowMode.

Fallout from 9f749767ac233270ffc58e423e83f09511bc2434

https://bugzilla.gnome.org/show_bug.cgi?id=775562