item-manager: Don't look for collections in wait_for_changes_table
Collections are stored in tracker's database. Unlike local BaseItems,
they are not going to undergo atomic file updates that can possibly
change their URN. Therefore we don't need to track them in the
wait_for_changes_table GHashTable.

Moreoever, local collections don't have a URI. So we don't have a key
to look for them in the GHashTable.

https://bugzilla.gnome.org/show_bug.cgi?id=776565