Move the code to update the OffsetControllers' count
Once we port to GtkFlowBox, we will lose the ViewModel class.
Therefore, we need to find a new home for the code that resets the
OffsetControllers' count whenever BaseItems are added or removed to
the child BaseManagers inside ItemManager.

TrackerController is a good choice because it already uses ItemManager
and the mode-specific OffsetController.

The other option was to do it in ItemManager itself, but since some
OffsetControllers also use ItemManager it will create problems when the
singletons are initially created.

https://bugzilla.gnome.org/show_bug.cgi?id=690623