item-manager: Add mechanism for constraining item additions
Constraining the addition of new items was an important function of the
ViewModel class. We want to show content in steps of N items, sorted by
their mtime in descending order. Normally, this works because the step
and offset are embedded in our SPARQL queries. However, when reacting
to out-of-band "graph-updated" signals, we might end up showing too
manyitems unless the constraints are re-validated.

Once we port to GtkFlowBox, we will lose the ViewModel class.
Therefore, we need to find a new home for this code.

This commit adds some of the mechanisms needed to perform constrained
item additions in ItemManager. The constraints aren't actually turned
on. That will happen in a latter commit.

https://bugzilla.gnome.org/show_bug.cgi?id=690623