image-view: Implement double-buffering
There are drawbacks to directly blitting the GeglNode that was passed
to PhotosImageView.

GEGL processes the graph in smaller chunks. The GeglNode might not be
fully processed when we need to draw the widget, leading to rendering
artifacts. We made this bearable in commit
d2f4589f7747265786aed0da4046235f63f8e8d4 by not queuing a draw after
each chunk is computed. However, the problem was still there. A draw
may be queued for reasons completely beyond control. eg., the widget
can get resized.

Secondly, the zoom and positioning of the image were being updated for
every chunk. It is wasteful. More importantly, due to the chunked
processing, the output of the GeglNode might not be ready to match the
latest value of the bounding box.

Therefore, a separate graph, with a single gegl:buffer-source, is used
for rendering. The output of the GeglNode is copied to a new
GeglBuffer and fed into this graph. Changes to the GeglNode's bounding
box, caused by modifications to its graph's connectivity or
properties, are tracked, and so is the amount of area
that has been computed. We update our GeglBuffer, zoom and positioning
only when the entire bounding box has been computed.