image-view: Don't use a gegl:buffer-source for double-buffering
Instead of using a GEGL graph with a gegl:buffer-source in it, we can
directly use the GeglBuffer that we copied out of the GeglNode.
Sometimes, it does take a few fractions of a second to process the
gegl:buffer-source before it can be used. Probably due to overheads
inherent to processing a GEGL graph. By directly using the GeglBuffer
we can save that time.

It is not clear if this is going to have any negative side-effects, or
whether the time that we saved is going to show up elsewhere. That is
why this has been done as a separate commit. We can always go back to
using a gegl:buffer-source if things don't work out.