base-item: Modify and process the pipeline in batches
There is a drawback to modifying the GEGL graph and processing it as
soon as photos_base_item_operation_{add, remove} are called.

GEGL might already be processing it in chunks in an idle handler.
Interrupting it will lead to GeglNode::invalidated being emitted,
which will mislead PhotosImageView into thinking that processing has
been restarted. If this happens too often, with a high enough number
of pixels or a graph that is noticeably slow, the view won't get
updated for a long time since it will keep waiting for the processing
to complete.

Therefore, instead of interrupting the graph we buffer all pending
edits. Once the previous run is over, we apply everything that's been
queued and process it afresh. This ensures that the view gets a chance
to update itself after every batch.