application: Try to speed up Babl conversions on the first run
Profiling the various pixel conversion code paths for a given source
and destination format might take enough time to make a visible
difference in some cases. For example, when sliding out the previews
of the magic filters. Newer versions of Babl keep the outcome of the
profiling in a persistent cache. Hence, it is only a problem when the
conversions are used for the first time after installing a new Babl
version.

This problem can be further reduced by populating Babl's cache before
a performance-critical operation takes place.

This is done in an idle handler after the main window is created so
that the user interface continues to show up as fast as possible. The
assumption being that just showing the main window doesn't require any
performance-critical conversions, and the time spent profiling won't be
noticeable if it is done right after the window is shown.