Add PhotosThumbnailFactory
This is the application-facing interface to the thumbnailer process.
It is modelled on gnome_desktop_thumbnail_factory_generate_thumbnail.
It adds error-handling and the generated thumbnail is implicitly saved
without handing the pixels back to the application.

There is a placeholder for cancelling running jobs, but it isn't
implemented at the moment.

https://bugzilla.gnome.org/show_bug.cgi?id=690255