Add an asynchronous version of gdk_pixbuf_new_from_file_at_size
The underlying gdk_pixbuf_new_from_file_at_size is not cancellable,
which is problematic, but will do for the time being. This should be
added to gdk-pixbuf proper and the lack of cancellation addressed.

We can't put this code in photos-utils.[ch] because we will need to use
it from the thumbnailer. There are too many build dependencies
associated with photos-utils.[ch] that we can't have in the
thumbnailer.