utils: Add photos_utils_file_copy_as_thumbnail
This is meant to copy a pre-existing thumbnail image to its proper
location with the necessary metadata and in the right format (ie. PNG).
Remote items often have server-side thumbnails, but won't have the
metadata that we need, or might not be in the format that we expect.
A simple g_file_copy will copy a JPEG image to a thumbnail file with
the '*.png' extension. That's not good. It can trigger strange corner
cases in various MIME-type detection code. Now that we are changing
the location of our thumbnail cache, it is a good time to address this
and avoiding a whole class of bugs.