source-manager: Keep track of sources that need attention
Note that SourceManager is shared between the application and the
search provider, and it is created before gtk_init is called. This
means that it can't use GTK+ widgets. Therefore we introduce one level
of indirection, based on two signals, so that the notifications can be
created from some other part of the code that is specific to the
application and is able to use GTK+.

Based on code written by Umang Jain.

https://bugzilla.gnome.org/show_bug.cgi?id=777492