thumbnailer-dbus: Add a Cancel method to support cancellation
Updating the thumbnails to reflect the edited state of the items
requires that ongoing thumbnailing jobs be cancelled on the
client-side. It will be nice to avoid any needless I/O or CPU
consumption by propagating the cancellation to the thumbnailer process.

This Cancel method is meant to be used by ThumbnailerFactory to tell
the thumbnailer to cancel a specific GenerateThumbnail call.

https://bugzilla.gnome.org/show_bug.cgi?id=763329