base-item: Delete the thumbnail if create_thumbnail was cancelled
In a subsequent patch, any ongoing thumbnailing job will be cancelled
when the pipeline is saved so that a new thumbnail can be created to
reflect the edited state of the item. We want to ensure that there are
no remnants from the cancelled operation before submitting a new one.
Since all thumbnailing operations are serialized in a single-threaded
GThreadPool, this will guarantee that any leftover has been cleaned up
before the new operation commences.

https://bugzilla.gnome.org/show_bug.cgi?id=763329