base-item: Shorten the lifetime of the loading graph
There is no need to keep the graph that loaded the GeglBuffer as an
instance variable for the entire lifetime of the BaseItem. It can be
entirely local to the function that does the loading.

This reduces the memory footprint of each BaseItem by three pointers.
Plus, any GeglBuffer that might be left loitering around in the graph.