base-item: Clean up the leftovers when _base_item_load_async fails
This prevents the BaseItem from being left in an inconsistent state.
For example, if the photos_base_item_load_async failure was due to a
cancellation, then subsequent attempts to load the BaseItem might
encounter a valid Pipeline and conclude that it is already loaded.
When, in fact, the buffer source might not have a GeglBuffer due to
the cancellation.

https://bugzilla.gnome.org/show_bug.cgi?id=777867