base-item: Check edit_graph to determine if the item is already loaded
In a subsequent commit photos_base_item_pipeline_is_edited_async will
be changed to only load the pipeline, and not the entire item. This
will make it much faster for RAW images. Therefore, we can no longer
use the availability of the pipeline to determine if an item is already
loaded. We now use priv->edit_graph instead.