spinner-box: Don't set any extra properties in g_object_new
The GObject recommendation is that the C convenience wrapper,
foo_bar_new (), should be same as g_object_new (FOO_TYPE_BAR, NULL).
In this case, it ensures that there is no difference between a
SpinnerBox that's created using the C convenience wrapper and one
created via a GtkBuilder template.

https://bugzilla.gnome.org/show_bug.cgi?id=780424