application, help-overlay, preview-view: Hook up zooming to GActions
There is no upper limit for the zoom, but we use the best-fit zoom
level as the lower limit. Best-fit is defined as the size of the image
in device pixels or the ImageView's allocation in device pixels,
whichever is smaller. See commit ccf511c2d6e2dd90 for a longer
explanation.

The initial zoom-in is much bigger compared to subsequent increments,
but zooming out always uses the smaller increments.

All the above policy is implemented in PreviewView, outside the
ImageView widget. Currently, we don't update the best-fit zoom factor
value when the window is resized. Users with particularly large
displays will notice that increasing the window size lets them
zoom-out beyond the best-fit value corresponding to the bigger window.
Similarly shrinking the window might skip the last few zoom-out steps
and jump directly to the best-fit level. However, the zoom increments
seem large enough for this to not be an issue in practice. Even if
someone does encounter this oddity, it is unlikely to be a real
annoyance.

https://bugzilla.gnome.org/show_bug.cgi?id=742662