tool-crop: Restrain the selection when changing aspect ratios
If the existing crop area is too close to the boundaries of the image,
then changing the aspect ratio can make it exceed the image. For
example, if we had a square selection that touched one edge, or covered
the entire width of a portrait, then changing the aspect ratio would
make it exceed the horizontal limits of the image.

In such cases, it is more important to ensure that the selection fits
inside the image than to retain its centre and area.

https://bugzilla.gnome.org/show_bug.cgi?id=765136