image-view, tool-crop: Use double instead of float in the public API
Even though we don't need the extra accuracy of double-precision
floating-point numbers, it doesn't cost us that much either because we
are not going to have hundreds of PhotosImageView instances. On the
other hand it does give us consistency with the Cairo and GVariant
APIs, and the future GtkImageView API that's in the works.