image-view: Implement smooth zooming
This is done by animating between the currently visible zoom level and
the new requested value. Since these intermediate zoom levels are not
meant to be exposed to users of the PhotosImageView widget, a small
helper type called PhotosImageViewHelper is used instead. The "zoom"
property of this helper is driven with EggAnimation, and the resulting
"notify" emissions are used by the widget for rendering.

https://bugzilla.gnome.org/show_bug.cgi?id=742662