image-view: Fix the range of the "zoom" property
The zoom factor of an image may approach 0.0, but it will always be a
positive value. So, the lowest acceptable value should be the minimum
positive value that can be held in a gdouble.

Since PhotosImageView doesn't implement any policy about the maximum
and minimum thresholds of the zoom factor, let's use the maximum value
that can be held in a gdouble as the upper limit.