application, preview-view: Tweak the zoom-in/out GActions for scrolling
The subsequent commit is going to use scroll events for zooming. Scroll
events have non-integer deltas, and users tend to scroll more as
as compared to repeatedly pressing the accelerator keys.

Therefore, the zoom-in/out GActions have been adjusted to work with
non-integer deltas. Two ticks on the mouse's scroll wheel has the same
effect as a single press of the corresponding accelerator key. This
differentiation is achieved by using -1.0 to denote a key press, and
non-negative deltas for the mouse and touch pad.

https://bugzilla.gnome.org/show_bug.cgi?id=742662