Revert "build: Update all AX_* macros from upstream"
This seems to have broken Continuous with:
  ../../src/photos-tracker-change-monitor.c:95:1: error: packed
    attribute is unnecessary for '_TrackerResourcesEvent'
    [-Werror=packed] } __attribute__ ((packed));

This warning and our use of -Wpacked aren't new. They have been present
for a while. It is not clear to me how Continuous managed to work so
far, and why it broke only now. I am also failing to spot any relevant
change in the macros that might have caused this.

This reverts commit 548c27eb80f806e94adfbf6061b897518f650aa4.