application: Instantiate SearchProvider only when registering on D-Bus
Now that the --version flag was implemented using the
handle_local_options virtual method, it is possible that the
Application may exit without ever touching D-Bus. So it is a tad
wasteful to instantiate the SearchProvider when it is never going to
be used. Since we are unreffing it in dbus_unregister, we might as well
create it in dbus_register.

This improves upon the fix in commit 8cfafad535ef83a0.