Use our own composer for the SVG (and PDF) blend operation multiply
The svg:multiply operation got removed in GEGL 0.3.16 due to conflicts
with gegl:multiply. Even though gegl:multiply is intended to produce
the same result as svg:multiply, that is not the case right now. They
only match when the auxiliary and input buffers don't have any
transparency.

Let's implement and carry our own until this is resolved.

Based on:
https://www.cairographics.org/operators/
https://www.w3.org/TR/2004/WD-SVG12-20041027/rendering.html#comp-op-prop