preview-view: Don't horizontally expand the EditPalette
The EditPalette should be just as wide as it needs to be, and shouldn't
needlessly expand. We have managed to survive so far by using GtkBoxes
to pack the tools' widgets, and because of the fact that their layouts
were relatively simple. However, that's no longer enough as the
ToolCrop grows more bells and whistles in the form of orientation
controls.

Therefore, let's do the right thing and set EditPalette's hexpand to
FALSE. As a side effect, it will be possible to use GtkGrids instead
of GtkBoxes for the tools.

https://bugzilla.gnome.org/show_bug.cgi?id=781134