tool-crop: Always express the "original" constraint as landscape
All the other constraints, except square, are always expressed as
landscapes regardless of the orientation of the original BaseItem.
This inconsistency is exposed by portrait BaseItems, where the
orientation changes from landscape to portrait depending on the chosen
constraint.

A subsequent commit will allow changing the orientation at will. Always
expressing the "original" constraint as landscape will help with that.
Cropping portraits was already painful, so this doesn't make it much
worse.

https://bugzilla.gnome.org/show_bug.cgi?id=781134