tool-crop: Clear "reset" earlier in the activation sequence
A subsequent commit will add support for changing the orientation of
the crop rectangle. For BaseItems without any pre-existing gegl:crop,
it will set the orientation of the crop to match that of the original
item. However, it is not allowed to change the crop rectangle if the
"reset" variable is set.

Therefore, the variable is now cleared earlier.

During the reset sequence, we update "activated" and "reset" in
separate stages. Therefore it seems to make sense to clear "reset" as
soon as the "activate" virtual method is called.

https://bugzilla.gnome.org/show_bug.cgi?id=781134