base-item: Use QUERY_FLAGS_UNFILTERED in base_item_refresh
The difference between QUERY_FLAGS_NONE and QUERY_FLAGS_UNFILTERED is
that the former applies the default WHERE and FILTER clauses (ie.
membership in the active collection, if any, and "all" for everything
else), while the latter doesn't.

Therefore, QUERY_FLAGS_NONE is appropriate for showing the contents of
a collection, or to verify if a newly created URN is relevant to the
application depending on the active collection, if any. However,
base_item_refresh is used with items that are already known, so their
relevance has already been established. In fact, QUERY_FLAGS_NONE
might mistakenly filter them out because they don't belong to the
active collection, even though they are still used by some other mode.