flatpak: Disable Tracker's filesystem miner
Like all the other miners shipped by Tracker, the filesystem miner is
meant to run on the host operating system. They don't add any value to
the Flatpak.

A few other miners and the Nautilus extension, which were not being
built so far due to missing dependencies, were explicitly disabled to
avoid any future accidents. The --disable-artwork option was removed
because it is redundant when --disable-tracker-needle is specified.