Replace gnome-desktop with gsettings-desktop-schemas
Since the switch to a custom thumbnailing infrastructure in commit
f8ceb9d27184ecd0, the only use for gnome-desktop were the
GDesktopBackgroundShading and GDesktopBackgroundStyle enumerated types.
Since those types are actually exported by gsettings-desktop-schemas,
it is better to directly use that instead of a separate intermediary.

This has the nice side-effect of reducing the size of the Flatpak.