operation-svg-multiply: Silence -Wimplicit-function-declaration, etc.
GeglOperation sub-classes didn't provide autocleanups with GEGL
0.3.14, which is the version that is required by the gnome-3-24
branch. Those were added later. It is also problematic to carry the
autocleanups in GNOME Photos, because that will break the build with
newer GEGL versions.

Since it is not nice to bump the requirement in a stable branch, the
easiest option is to use the old-fashioned boilerplate instead.

(The other option was to use #ifdef's based on the GEGL version, but
that's needlessly messy.)

Fallout from f859aa71b4537d5a787d89689757a5ede602f7d4