embed: Simplify code
The purpose of the photos_embed_search_changed callback is to switch
to the SEARCH mode when the user specifies any search criterion, and
to switch back when she has lifted all criteria. It shouldn't interfere
when the application is saving and restoring search state for internal
state management.

Therefore, instead of a mysterious comment and hard to understand
conditions inside photos_embed_search_changed itself, it is simpler to
temporarily block it whenever there is a change due to internal
bookkeeping.

https://bugzilla.gnome.org/show_bug.cgi?id=786936