embed: Untangle the notify::visible-child & window-mode-changed logic
The purpose of the GtkStack::notify::visible-child handler is to change
the WindowMode when the GtkStackSwitcher is used. That's only possible
with a left-click or tapping the touchscreen.

The ModeController::window-mode-changed handler is meant to do all
that's necessary when the mode changes - setting up the widgets,
saving application state, etc..

So far, both the handlers would be invoked for every mode change
regardless of whether the GtkStackSwitcher was used or not; but the
order in which they ran was determined by the involvement of the
GtkStackSwitcher. That's confusing.

From now on, the photos_embed_notify_visible_child will ignore changes
to the visible child unless it happened due to a left-click or touch
event. The logic to determine whether to hide the searchbar or not has
been moved to photos_embed_window_mode_changed because it is just like
the rest of the window-mode-changed handling code, and has nothing to
do with the GtkStackSwitcher.

https://bugzilla.gnome.org/show_bug.cgi?id=786936