embed: Don't keep the searchbar active merely because it has the focus
The searchbar shouldn't get yanked out (or hidden) when the user is
actively changing the search constraints, but that's not the same as
having the focus. The searchbar might still be focused when the user
is using the GtkStackSwitcher to change the modes.

Fallout from 770366286115c2c8155ae9cb71a2033fc6f0cd05

https://bugzilla.gnome.org/show_bug.cgi?id=786936