Consolidate code to update MainToolbar on ItemManager::items-changed
Most GActions are enabled/disabled in photos_application_actions_update
when there is a state change. Now that the selection mode is backed by
a GAction, there is no reason to directly enable/disable the GtkButton
when the number of BaseItems change.

This also disables all the other app.search-* and app.selection-*
GActions when there are no BaseItems in the current mode.

https://bugzilla.gnome.org/show_bug.cgi?id=786936