Let PhotosQuery instances be reference counted
... by turning it into a GObject.

The other option was to implement a boxed type with its own reference
counting using g_atomic_*. However, there is no proof that using a
GObject will measurably affect performance and we get the reference
counting for free.

This paves the way for reduced memory fragmentation, by removing the
need to copy the sparql and tag fields of a PhotosQuery.