tool-crop: Make "constraints" an instance variable
The subsequent patch will hide the orientation controls for the
"original" aspect ratio if the item is a square. This will be done by
updating the orientable field of the ORIGINAL constraint in the
activated virtual method. Since it is not correct to modify a
class-wide variable from an instance, its scope needs to be restricted
to a single instance.

https://bugzilla.gnome.org/show_bug.cgi?id=785115